requirements_draft.txt
brainflow==4.2.0 #brainflow may need pip uninstall and pip install -U brainflow
hyperopt==0.2.7
matplotlib==3.1.3
numpy==1.17.0 #possibly numpy==1.8.1
pandas==1.0.3
Pillow==9.4.0
py7zr==0.16.1
pyserial==3.5 #possibly 3.4
#scikit_learn==1.2.1
scikit-learn==0.22.1
scipy==1.6.2
